requirements.txt
asgiref==3.9.2
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
defusedxml==0.7.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.11.2
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
gunicorn==23.0.0
hedera-sdk-py==2.50.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyjnius==1.7.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyotp==2.9.0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.3
pyzbar==0.1.9
qrcode==8.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
service-identity==24.2.0
setuptools==80.9.0
social-auth-app-django==5.5.1
social-auth-core==4.7.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
zope.interface==8.0.1
fido2
google-auth